vlm_benchmark.py
import torch
import time
import os
from PIL import Image
from transformers import Qwen2_5_VLForConditionalGeneration, AutoProcessor, BitsAndBytesConfig, AutoModelForCausalLM, AutoTokenizer
from qwen_vl_utils import process_vision_info

# Environment: Acer Nitro 5 (32GB RAM / GPU optimized)
# Measurement for Prefill and Decoding Latency

def benchmark_qwen(image_path):
    model_id = "Qwen/Qwen2.5-VL-7B-Instruct"
    bnb_config = BitsAndBytesConfig(
        load_in_4bit=True,
        bnb_4bit_compute_dtype=torch.float16,
        bnb_4bit_quant_type="nf4",
        bnb_4bit_use_double_quant=True
    )
    model = Qwen2_5_VLForConditionalGeneration.from_pretrained(
        model_id, torch_dtype=torch.float16, device_map="auto", quantization_config=bnb_config
    )
    processor = AutoProcessor.from_pretrained(model_id)
    
    # Benchmarking logic (Prefill & Decoding)
    # [Insert benchmarking logic here based on our successful runs]
    print("Qwen-7B Benchmark Complete")

def benchmark_moondream(image_path):
    model_id = "vikhyatk/moondream2"
    model = AutoModelForCausalLM.from_pretrained(model_id, trust_remote_code=True).to("cuda")
    # [Insert moondream logic here]
    print("Moondream2 Benchmark Complete")

if __name__ == "__main__":
    print("VLM Performance Environment Initialized")