requirements.txt
transformers>=4.49.0
accelerate
bitsandbytes
qwen-vl-utils
pillow
torch